[kotlin] Release v0.6.2
Remove unsigned integer from public interface

+++ clients/kotlin/src/main/kotlin/net/eternaltwin/dinoparc/DinoparcDinozElements.kt
@@ -8,11 +8,11 @@ import kotlinx.serialization.*
  */
 @Serializable
 data class DinoparcDinozElements constructor(
-  val fire: UShort,
-  val earth: UShort,
-  val water: UShort,
-  val thunder: UShort,
-  val air: UShort,
+  val fire: Int, // TODO: UShort
+  val earth: Int, // TODO: UShort
+  val water: Int, // TODO: UShort
+  val thunder: Int, // TODO: UShort
+  val air: Int, // TODO: UShort
 ) {
   companion object {
     @JvmStatic
